+++ crates/comptime/husky-rust-transpilation/expect-files/examples/mnist-classifier/module_defn_rust_transpilation/digits/four.rs
@@ -17,22 +17,22 @@ pub fn components_max_heights() -> FermiMatchResult {
 }
 
 pub fn is_four() -> OneVsAll {
-    require!(let some = left_components().matches[0]);
-    require!(let some = left_components().matches[1]);
+    require!(let some = left_components().matches[0 as usize]);
+    require!(let some = left_components().matches[1 as usize]);
     let eff_holes = major_connected_component().eff_holes();
-    require!(let none = eff_holes.matches[1]);
-    let down_match = components_max_downwards().matches[0];
+    require!(let none = eff_holes.matches[1 as usize]);
+    let down_match = components_max_downwards().matches[0 as usize];
     require!(let some = down_match);
     let down_match_dp_y = down_match.unwrap().displacement().y;
     let higher_excess = major_connected_component().upper_mass() - major_connected_component().lower_mass();
     require!(higher_excess > 7.0f32);
-    if let none = eff_holes.matches[0] {
+    if let none = eff_holes.matches[0 as usize] {
         require!(major_concave_components().ilen() >= 2);
-        let four_match_refine_result = components_max_heights().matches[0];
+        let four_match_refine_result = components_max_heights().matches[0 as usize];
         require!(let some = four_match_refine_result);
         require!(components_max_heights().norm() < 1.0f32);
         let higher_excess = major_connected_component().upper_mass() - major_connected_component().lower_mass();
-        let upper_arc = components_max_heights().matches[0];
+        let upper_arc = components_max_heights().matches[0 as usize];
         require!(let some = upper_arc);
         require!(upper_arc.unwrap().displacement().y > 0.0f32);
         require!(upper_arc.unwrap().angle_change() < -110.0f32);
