+++ examples/mnist-classifier/target-rs/mnist-classifier/src/digits/seven.rs
@@ -36,10 +36,10 @@ pub fn is_seven() -> OneVsAll {
     require!(major_connected_component().max_hole_ilen() == 0.0f32);
     let simple_match_norm = simple_seven_match().norm();
     if simple_match_norm < 1.0f32 {
-        require!(let some = simple_seven_match().matches[0]);
+        require!(let some = simple_seven_match().matches[0 as usize]);
         let upper_excess = major_connected_component().upper_mass() - major_connected_component().lower_mass();
         if upper_excess < 10.0f32 {
-            let end_tangent = simple_seven_match().matches[0].unwrap().end_tangent();
+            let end_tangent = simple_seven_match().matches[0 as usize].unwrap().end_tangent();
             let a = end_tangent.y;
             require!(a < -7.0f32);
         }
@@ -50,7 +50,7 @@ pub fn is_seven() -> OneVsAll {
         require!(upper_excess > 10.0f32);
         return OneVsAll::Yes;
     }
-    require!(let some = special_seven_match().matches[0]);
+    require!(let some = special_seven_match().matches[0 as usize]);
     let others = special_seven_match().others;
     require!(false);
     OneVsAll::Yes
