+++ examples/mnist-classifier/target-rs/mnist-classifier/src/major.rs
@@ -8,19 +8,19 @@ pub fn major_connected_component() -> Leash<ConnectedComponent> {
     let mut i0 = 0;
     let mut max_row_span_sum = 0.0f32;
     for i in 0..connected_components().ilen() {
-        let row_span_sum = connected_components()[i].row_span_sum();
+        let row_span_sum = connected_components()[i as usize].row_span_sum();
         if row_span_sum > max_row_span_sum {
             max_row_span_sum = row_span_sum;
             i0 = i
         }
     }
-    return connected_components()[i0];
+    return connected_components()[i0 as usize];
 }
 
 pub fn ignored_connected_components_row_span_sum_sum() -> f32 {
     let mut sum = 0.0f32;
     for i in 0..connected_components().ilen() {
-        sum += connected_components()[i].row_span_sum()
+        sum += connected_components()[i as usize].row_span_sum()
     }
     return sum - major_connected_component().row_span_sum();
 }
@@ -30,7 +30,7 @@ pub fn major_raw_contours() -> Leash<Vec<RawContour>> {
 }
 
 pub fn major_raw_contour() -> Leash<RawContour> {
-    &major_connected_component().raw_contours()[0]
+    &major_connected_component().raw_contours()[0 as usize]
 }
 
 pub fn major_line_segment_sketch() -> Leash<LineSegmentSketch> {
